Rename package gene_nalyzer to gene_analyzer
