+++ src/main/java/com/epam/gene_analyzer/Model/Area.java
@@ -1,4 +1,4 @@
-package com.epam.gene_nalyzer.Model;
+package com.epam.gene_analyzer.Model;
 
 import java.awt.*;
 import java.awt.image.BufferedImage;
