+++ src/main/java/com/epam/gene_analyzer/Window/CanvasGraph.java
@@ -1,7 +1,7 @@
-package com.epam.gene_nalyzer.Window;
+package com.epam.gene_analyzer.Window;
 
-import com.epam.gene_nalyzer.Controllers.MainController;
-import com.epam.gene_nalyzer.Model.Pixel;
+import com.epam.gene_analyzer.Controllers.MainController;
+import com.epam.gene_analyzer.Model.Pixel;
 
 import javax.swing.*;
 import java.awt.*;
