+++ src/main/java/com/epam/gene_analyzer/Window/Menu.java
@@ -1,6 +1,6 @@
-package com.epam.gene_nalyzer.Window;
+package com.epam.gene_analyzer.Window;
 
-import com.epam.gene_nalyzer.Controllers.MainController;
+import com.epam.gene_analyzer.Controllers.MainController;
 
 import javax.swing.*;
 import java.awt.*;
