+++ src/main/java/com/epam/gene_analyzer/Window/TableWindow.java
@@ -1,7 +1,7 @@
-package com.epam.gene_nalyzer.Window;
+package com.epam.gene_analyzer.Window;
 
-import com.epam.gene_nalyzer.Controllers.MainController;
-import com.epam.gene_nalyzer.Model.RtModel;
+import com.epam.gene_analyzer.Controllers.MainController;
+import com.epam.gene_analyzer.Model.RtModel;
 
 import javax.swing.*;
 import javax.swing.event.TableModelEvent;
